Use same casing in wrapper and wrapped classes
Use the same casing for wrapper classes attribute names as is used for
the respective wrapped class attribute names.  That is, change the
casing of the names of `CompendiumEntryData`'s from PascalCasing to
camelCasing so as to match the casing of the names of
`CompendiumEntry`'s attributes, even though this violates C# conventions
for public attribute naming.