Change keys used when serializing LocalizedString
When serializing a `LocalizedString` with
`LocalizedStringConverter<T>.WriteObjectBody(...)` method, use
`locale.Identifier.Code` rather than `locale.ToString()` as the key/name
for the localized value for each `locale`.

This change is made because, with the latest version of Lost Skies (full
release, v. 1.0.79), the value produced by `locale.ToString()` changed
in a manner that has made it less convenient to for keys/names in
serialization.  The values produced by passing the result of value
produced by `locale.ToString()` to
`EncodeName(string, JsonSerializerOptions)` when using the mod with the
previous (early access) and current versions of Lost Skies are tabulated
below.  Note that the Korean locale was available in the previous
version of Lost Skies, but is not available in the current version.

| Prior Encoded Value            | Current Encoded Value                        |
|--------------------------------|----------------------------------------------|
| `Chinese (Simplified) (zh-CN)` | `\u7B80\u4F53\u4E2D\u6587`                   |
| `English (en-GB)`              | `English`                                    |
| `French (fr-FR)`               | `Fran\u00E7ais`                              |
| `German (de-DE)`               | `Deutsch`                                    |
| `Italian (it-IT)`              | `Italiano`                                   |
| `Japanese (ja-JP)`             | `\u65E5\u672C\u8A9E`                         |
| `Korean (ko-KR)`               |                                              |
| `LatinAmerica (es-LA)`         | `Espa\u00F1ol (Latinoam\u00E9rica)`          |
| `Portuguese (Brazil) (pt-BR)`  | `Portugu\u00EAs (Brasil)`                    |
| `Russian (Russia) (ru-RU)`     | `\u0420\u0443\u0441\u0441\u043A\u0438\u0439` |
| `Spanish (es-ES)`              | `Espa\u00F1ol (Espa\u00F1a)`                 |